Move relational expressions into expr.relational_expr
